moved constants to TorrentDefaults class
