--- src/main/java/com/turn/ttorrent/client/network/keyProcessors/Constants.java
@@ -1,7 +1,0 @@
-package com.turn.ttorrent.client.network.keyProcessors;
-
-public class Constants {
-
-  public final static int SOCKET_CONNECTION_TIMEOUT_MILLIS = 7000;
-
-}
